Move _DiskArray out of tests
