+++ Project.toml
@@ -8,10 +8,10 @@ OffsetArrays = "6fe1bfb0-de20-5000-8ca7-80f57d26f881"
 
 [compat]
 Aqua = "0.8"
-OffsetArrays = "1"
-Statistics = "1.6"
-Test = "1.6"
 julia = "1.6"
+OffsetArrays = "1"
+Statistics = "1.9"
+Test = "1.9"
 
 [extras]
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
